third_party/WebKit/LayoutTests/platform/win/webaudio/unit-tests/testharness-test-expected.txt

This is a testharness.js-based test.
PASS Zero is equal to 0. 
PASS One is not equal to 0. 
PASS Expected SNR (110) is greater than or equal to 100. 
PASS Maximum error value (0.000001) is less than or equal to 0.00001. 
PASS One point double zero one is 1 within a relative error of 0.1. 
PASS [2, 2, 2] contains only the constant 2. 
PASS [1, 2, 3] is identical to the array [1,2,3]. 
PASS My array equals [0.1,0.2] with an element-wise tolerance of 0.02. 
PASS My random array contains all the expected values in the correct order: [1,3,2]. 
Harness: the test ran to completion.
